:truck: Rename test/helpers to test/fixtures
