+++ test/cookie-session-koa-test.js
@@ -4,7 +4,7 @@ var request = require('supertest');
 var should = require('should');
 
 const AV = require('..');
-const appInfo = require('./helpers/app-info');
+const appInfo = require('./fixtures/app-info');
 
 var app = koa();
