+++ test/fixtures/app-info.js
