+++ test/fixtures/frameworks.js
@@ -1,9 +1,8 @@
 const koa = require('koa');
-const request = require('supertest');
 
 const AV = require('../..');
 
-exports.app = function() {
+module.exports = function() {
   if (process.env.FRAMEWORK == 'koa') {
    var app = koa();
    app.use(AV.koa());
