+++ test/fixtures/functions.js
